Added Command Processing & some other functions too.
Added Commands Processing. It allows one to enter commands on the
command line or store them into some file to make different graphical
objects and make transformations to them.

Code is cluttered heavily, but I'm focusing on CommandProcessor as of
now and will clean up wherever necessary in free time.